requirements.txt
asgiref==3.8.1
astroid==3.2.2
black==24.3.0
click==8.1.7
dill==0.3.8
Django==5.0
django-browser-reload==1.12.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.2.2
pytest-django==4.7.0
python-dotenv==1.0.1
ruff==0.2.1
setuptools==69.5.1
sqlparse==0.5.0
tomlkit==0.12.5
wheel==0.43.0
